requirements.txt
Django>=4.2,<5.0
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
channels>=4.0.0
channels-redis>=4.1.0
django-environ>=0.11.0
djangorestframework>=3.14.0
django-filter>=23.0.0
psycopg2-binary>=2.9.0
psycopg2>=2.9.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
twilio>=8.10.0
python-dotenv>=1.0.0
kombu>=5.3.0
redis>=5.0.0
daphne>=4.0.0
pytest>=7.0.0
pytest-django>=4.5.0
coverage>=7.0.0
django-modeltranslation>=0.18.11
django-rosetta>=0.9.9
django-extensions>=3.2.3
django-waffle>=3.0.0
